+++ client/components/Application/Dialogs/DeleteFile/index.ts
@@ -1,0 +1,1 @@
+export { DeleteFileDialog } from './DeleteFile'
