+++ client/types/index.ts
@@ -1,17 +1,18 @@
 export * from './column'
 export * from './config'
 export * from './contributor'
+export * from './dialect'
 export * from './error'
 export * from './event'
 export * from './file'
 export * from './general'
 export * from './help'
 export * from './history'
-export * from './dialect'
 export * from './license'
+export * from './menu'
 export * from './package'
 export * from './portal'
-export * from './menu'
+export * from './progress'
 export * from './record'
 export * from './report'
 export * from './resource'
