+++ client/types/progress.ts
@@ -1,0 +1,6 @@
+export type IProgress = {
+  type: 'error' | string
+  title?: string
+  message?: string
+  blocking?: boolean
+}
